IMGTOGHIBLI AI/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0
flask>=2.3.0
flask-cors>=3.0.10
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0
safetensors>=0.3.1
moviepy>=1.0.3
tqdm>=4.65.0
python-dotenv>=1.0.0
huggingface_hub>=0.16.0
